Add Apache2 conf example
I'm not sure if this is 100% accurate but I got help setting up an Apache2 vhost from a Discord member and thought it would be useful to take what they helped me with and use it as an example in the docs. Consider it a starting point for this, I'm sure a revision might be needed.